Refactor and clean codebase to enhance readability and maintainability.
- Address field ordering inconsistencies across several classes and regions.
- Update instance initialization and parameter handling to clarify intent and reduce redundancy.
- Improve shader parameter configuration for consistent assignments and remove unused/duplicated logic.
- Reorganize partial methods to centralize their definitions and streamline file structures.
- Adjust `GamePacketHandler` methods and event publishing for cohesive packet handling.
- Tweak log formatting for better readability and adherence to existing repository standards.

+++ 04.Client.Core/MartialHeroes.Client.Domain.Quests/Quests/QuestRecord.cs
@@ -76,7 +76,7 @@ public readonly struct QuestRecord
 
     public byte AcceptedFlag(int index)
     {
-        return (uint)index < (uint)AcceptedFlagsBuffer.Length ? AcceptedFlags[index] : (byte)0;
+        return (uint)index < AcceptedFlagsBuffer.Length ? AcceptedFlags[index] : (byte)0;
     }
 
     [InlineArray(6)]
